requirements.txt
Flask==3.0.0
PyJWT==2.8.0
cryptography==41.0.7
bcrypt==4.1.2